Error in line 92
Needs to be a different (new) file that's just the seq names. The current file lists all job submit information. 
I created a new file in the pre_pre_processing.sh script  (#all_seqs.list) that lists all the file names instead of calling on the location of each file name in the list entry of the swarm submission file (which would totally work but the other way is just simpler to me).
...this is technically more of an edit to the other file, but we started here, so here we are, together.